Fixed sms not sent bug.
For some reason sms were not delivered. The previous code works before. As a workaround I commented the line which puts the system to sleep and deletes the event created. You can just hide the specific calendar on your Calendar. That way there's no messy notification events. This is just a temporary fix until I find a workaround to delete events w/o sacrificing sms notification; possibly a function that scrapes for previous events and deletes it. CHEERS!